Issue 664 Create BLAST database when reference is registered

+++ server/ngb-cli/src/main/java/com/epam/ngb/cli/entity/blast/CreateDatabaseRequest.java
@@ -35,6 +35,6 @@ public class CreateDatabaseRequest implements RequestPayload {
     private String pathToFile;
     private Boolean parseSeqIds;
     private Integer blastDbVersion;
-    private Integer taxId;
+    private Long taxId;
     private String dbType;
 }
